// frontend/src/api/types.ts
export type RoyalLevel = 'None' | 'R1' | 'R2' | 'R3' | 'R4' | 'R5' | 'R6';
export type PrimeLevel = 'None' | 'P1' | 'P2' | 'P3';
// ...
export interface AchievementsDto {
  topGifterTimes: number;
  topReceiverTimes: number;
  celebrityOfTheMonthTimes: number;
  weeklyTopClubTimes: number;
}

export interface StatsDto {
  level: number;
  activeSeconds: number;
  clubsFollowed: number;
  clubsJoined: number;
  giftsSent: number;
  giftsReceived: number;
  profileViews: number;
}